renamed package name to com.chongzixin.locationstalker
