+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 29
 
     defaultConfig {
-        applicationId "com.google.android.gms.location.sample.locationupdatesforegroundservice"
+        applicationId "com.chongzixin.locationstalker"
         minSdkVersion 16
         targetSdkVersion 29
         versionCode 1
