+++ app/src/main/java/com/chongzixin/locationstalker/StartupBootReceiver.java
@@ -1,4 +1,4 @@
-package com.google.android.gms.location.sample.locationupdatesforegroundservice;
+package com.chongzixin.locationstalker;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -9,7 +9,7 @@ import android.util.Log;
 public class StartupBootReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
-        Log.i("StartupBootReceiver", "boot completed from com.google.android.gms.location.sample.locationupdatesforegroundservice");
+        Log.i("StartupBootReceiver", "boot completed from com.google.locationupdatesforegroundservice");
 
         Intent serviceIntent = new Intent(context, LocationUpdatesService.class);
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
